ci: validate commit messages
I added commitlint, Husky, and commitizen as NPM dependencies to the
product. commitlint will be used to validate that commit messages are
properly formatted and written correctly. Husky will be used to run
commitlint before every local commit using the commit-msg Git hook to
ensure that only valid commit messages are persisted to the repository.
commitizen is being used to help developers to write proper commit
messages and format them correctly to be persisted to the repository.

I created a GitHub Actions workflow that will run commitlint to validate
commit messages for all commits in a pull request and for the latest
commit that is merged into the main branch. This will ensure that we are
enforcing commit message rules for the product.